Make declaration match.
git-svn-id: svn://openpgp.nominet.org.uk/openpgpsdk/trunk@405 228a932f-eeec-0310-9813-fb354622df21

+++ include/openpgpsdk/packet-parse.h
@@ -129,7 +129,7 @@ enum ops_parse_type_t
 void ops_parse_options(ops_parse_info_t *pinfo,ops_content_tag_t tag,
 		       ops_parse_type_t type);
 
-ops_boolean_t ops_limited_read(unsigned char *dest,unsigned length,
+ops_boolean_t ops_limited_read(unsigned char *dest,size_t length,
 			       ops_region_t *region,ops_error_t **errors,
 			       ops_reader_info_t *rinfo,
 			       ops_parse_cb_info_t *cbinfo);
